--- tests/unit/utils/make-helper-test.js
@@ -1,9 +1,0 @@
-import makeHelper from 'ember-svg-jar/utils/make-helper';
-import { module, test } from 'qunit';
-
-module('Unit | Utility | make helper', function () {
-  test('it works', function (assert) {
-    let result = makeHelper();
-    assert.ok(result);
-  });
-});
